updated `/stats help`, more error msgs for `/tp:save`, refactoring for Pair.java
