+++ rvas-core_v0.3/src/main/java/core/commands/restricted/Speeds.java
@@ -1,6 +1,6 @@
 package core.commands.restricted;
 
-import core.backend.Pair;
+import core.data.objects.Pair;
 import core.events.SpeedLimiter;
 
 import java.util.*;
