+++ rvas-core_v0.3/src/main/java/core/data/objects/Pair.java
@@ -1,4 +1,4 @@
-package core.backend;
+package core.data.objects;
 
 public class Pair<L,R> {
 	private final L left; private final R right;
